+++ example/v1/get-user.ts
@@ -10,7 +10,7 @@ enum Status {
 export const getUserEndpoint = endpointsFactory
   .addMiddleware(methodProviderMiddleware)
   .build({
-    methods: ['get'],
+    method: 'get',
     description: 'example user retrieval endpoint',
     input: z.object({
       id: z.string().transform((id) => parseInt(id, 10))