+++ src/endpoint.ts
@@ -1,8 +1,9 @@
+import {Request, Response} from 'express';
 import {Logger} from 'winston';
 import {z} from 'zod';
 import {ConfigType} from './config-type';
 import {combineEndpointAndMiddlewareInputSchemas, getInitialInput, IOSchema, Merge} from './helpers';
-import {Request, Response} from 'express';
+import {Method, MethodsDefinition} from './method';
 import {MiddlewareDefinition} from './middleware';
 import {defaultResultHandler, ResultHandler} from './result-handler';
 
@@ -13,7 +14,6 @@ export type Handler<IN, OUT, OPT> = (params: {
 }) => Promise<OUT>;
 
 export abstract class AbstractEndpoint {
-  protected methods: Method[] = [];
   protected description?: string;
 
   public abstract execute(params: {
@@ -23,56 +23,63 @@ export abstract class AbstractEndpoint {
     config: ConfigType
   }): Promise<void>;
 
-  public getMethods() {
-    return this.methods;
-  }
-
   public getDescription() {
     return this.description;
   }
 
+  abstract getMethods(): Method[];
   abstract getInputSchema(): IOSchema;
   abstract getOutputSchema(): IOSchema;
 }
 
-export type Method = 'get' | 'post' | 'put' | 'delete' | 'patch';
+export type EndpointInput<T> = T extends Endpoint<infer IN, any, infer mIN, any, any> ? z.input<Merge<IN, mIN>> : never;
 
-export type EndpointInput<T> = T extends Endpoint<infer IN, any, infer mIN, any> ? z.input<Merge<IN, mIN>> : never;
+export type EndpointOutput<T> = T extends Endpoint<any, infer OUT, any, any, any> ? z.output<OUT> : never;
 
-export type EndpointOutput<T> = T extends Endpoint<any, infer OUT, any, any> ? z.output<OUT> : never;
+type EndpointProps<IN extends IOSchema, OUT extends IOSchema, mIN, OPT, M extends Method> = {
+  middlewares: MiddlewareDefinition<any, any, any>[];
+  inputSchema: IN;
+  outputSchema: OUT;
+  handler: Handler<z.output<Merge<IN, mIN>>, z.input<OUT>, OPT>;
+  resultHandler: ResultHandler | null;
+  description?: string;
+} & MethodsDefinition<M>;
 
 /** mIN, OPT - from Middlewares */
-export class Endpoint<IN extends IOSchema, OUT extends IOSchema, mIN, OPT> extends AbstractEndpoint {
+export class Endpoint<IN extends IOSchema, OUT extends IOSchema, mIN, OPT, M extends Method> extends AbstractEndpoint {
+  protected methods: M[] = [];
   protected middlewares: MiddlewareDefinition<any, any, any>[] = [];
   protected inputSchema: Merge<IN, mIN>; // combined with middlewares input
   protected outputSchema: OUT;
   protected handler: Handler<z.output<Merge<IN, mIN>>, z.input<OUT>, OPT>
   protected resultHandler: ResultHandler | null;
 
-  constructor({methods, middlewares, inputSchema, outputSchema, handler, resultHandler, description}: {
-    methods: Method[];
-    middlewares: MiddlewareDefinition<any, any, any>[],
-    inputSchema: IN,
-    outputSchema: OUT,
-    handler: Handler<z.output<Merge<IN, mIN>>, z.input<OUT>, OPT>
-    resultHandler: ResultHandler | null,
-    description?: string
-  }) {
+  constructor({
+    middlewares, inputSchema, outputSchema, handler, resultHandler, description, ...rest
+  }: EndpointProps<IN, OUT, mIN, OPT, M>) {
     super();
-    this.methods = methods;
     this.middlewares = middlewares;
     this.inputSchema = combineEndpointAndMiddlewareInputSchemas<IN, mIN>(inputSchema, middlewares);
     this.outputSchema = outputSchema;
     this.handler = handler;
     this.resultHandler = resultHandler;
     this.description = description;
+    if ('methods' in rest) {
+      this.methods = rest.methods;
+    } else {
+      this.methods = [rest.method];
+    }
   }
 
-  public getInputSchema(): IOSchema {
+  public getMethods(): M[] {
+    return this.methods;
+  }
+
+  public getInputSchema(): Merge<IN, mIN> {
     return this.inputSchema;
   }
 
-  public getOutputSchema(): IOSchema {
+  public getOutputSchema(): OUT {
     return this.outputSchema;
   }