+++ src/errors.ts
@@ -1,0 +1,5 @@
+export class DependsOnMethodError extends Error {
+}
+
+export class OpenAPIError extends Error {
+}