+++ src/index.ts
@@ -1,13 +1,15 @@
 export { ConfigType, LoggerConfig } from './config-type';
-export { AbstractEndpoint, Method, EndpointInput, EndpointOutput } from './endpoint';
+export { AbstractEndpoint, EndpointInput, EndpointOutput } from './endpoint';
+export { Method } from './method';
 export { EndpointsFactory } from './endpoints-factory';
 export { IOSchema, FlatObject } from './helpers';
 export { createLogger } from './logger';
 export { createMiddleware } from './middleware';
 export { ResultHandler } from './result-handler';
-export { Routing } from './routing';
+export { Routing, DependsOnMethod } from './routing';
 export { createServer, attachRouting } from './server';
 export { OpenAPI } from './open-api';
+export { OpenAPIError, DependsOnMethodError } from './errors';
 
 import { z } from 'zod';
 import * as createHttpError from 'http-errors';