+++ src/method.ts
@@ -1,0 +1,7 @@
+export type Method = 'get' | 'post' | 'put' | 'delete' | 'patch';
+
+export type MethodsDefinition<M extends Method> = {
+  methods: M[];
+} | {
+  method: M;
+};